+++ src/cal3d/coresubmesh.cpp
@@ -125,13 +125,13 @@ void CalCoreSubmesh::addVertex(const Vertex& vertex, CalColor32 vertexColor, con
     m_influences.insert(m_influences.end(), inf.begin(), inf.end());
 }
 
-int CalCoreSubmesh::addCoreSubMorphTarget(boost::shared_ptr<CalCoreSubMorphTarget> pCoreSubMorphTarget) {
-    int subMorphTargetId = m_vectorCoreSubMorphTarget.size();
+size_t CalCoreSubmesh::addCoreSubMorphTarget(boost::shared_ptr<CalCoreSubMorphTarget> pCoreSubMorphTarget) {
+    size_t subMorphTargetId = m_vectorCoreSubMorphTarget.size();
     m_vectorCoreSubMorphTarget.push_back(pCoreSubMorphTarget);
     return subMorphTargetId;
 }
 
-int CalCoreSubmesh::getCoreSubMorphTargetCount() {
+size_t CalCoreSubmesh::getCoreSubMorphTargetCount() {
     return m_vectorCoreSubMorphTarget.size();
 }
